+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `buildDateTimeColumn` helper
+
 ## [1.7.1] - 2021-03-22
 ### Changed
 - Update module name & description